feat: load documents on vector table

+++ src/application/usecase/remove-file.ts
@@ -1,13 +1,23 @@
 import { Injectable } from '@nestjs/common';
+import { GraphQLError } from 'graphql';
 import { KnexFileRepository } from 'src/infra/repository/knex.file.repository';
+import { KnexVectorTableRepository } from 'src/infra/repository/knex.vector.table.repository';
 
 @Injectable()
 export class RemoveFileUsecase {
-  constructor(private readonly fileRepository: KnexFileRepository) {
+  constructor(
+    private readonly fileRepository: KnexFileRepository,
+    private readonly vectorTableRepository: KnexVectorTableRepository,
+  ) {
     this.fileRepository = fileRepository;
+    this.vectorTableRepository = vectorTableRepository;
   }
 
   async execute(id: string) {
+    const file = await this.fileRepository.findById(id);
+    if (!file) throw new GraphQLError('File not found');
+
+    await this.vectorTableRepository.removeAllByFileId(id);
     return this.fileRepository.remove(id);
   }
 }
